requirements.txt
# GAIA-APP Python Requirements
# Install: pip install -r requirements.txt

# ── Core API ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# ── Web Scraping ──────────────────────────────────────────────────────
trafilatura>=1.8.0          # Best-in-class main-content extractor
beautifulsoup4>=4.12.0      # Fallback HTML parser
lxml>=5.0.0                 # Fast HTML/XML parsing backend

# ── LLM Providers (install whichever you use) ─────────────────────────
openai>=1.20.0              # OpenAI GPT-4o-mini (set OPENAI_API_KEY)
anthropic>=0.25.0           # Anthropic Claude   (set ANTHROPIC_API_KEY)
# ollama                    # Ollama local models (set OLLAMA_MODEL=mistral)

# ── Search Providers (all optional — DDG works with no key) ───────────
# brave-search              # Brave Search API   (set BRAVE_API_KEY)
# tavily-python             # Tavily AI Search   (set TAVILY_API_KEY)

# ── Optional: Vector Memory (RAG) ─────────────────────────────────────
# chromadb>=0.4.0           # Local vector store for long-term RAG memory
# sentence-transformers     # Embeddings for ChromaDB

# ── Earth Intelligence (ATLAS layer) ──────────────────────────────────
# earthengine-api>=0.1.390  # Google Earth Engine (run: earthengine authenticate)
